feat(api): Adding sevice to call spotify api chore(helper): Moving getRandomObject to utils directory
